finish phasing out persistentdata class
