+++ src/main/java/factionsplusplus/builders/MessageBuilder.java
@@ -1,9 +1,12 @@
 package factionsplusplus.builders;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
-public class MessageBuilder {
+import factionsplusplus.builders.interfaces.GenericMessageBuilder;
+
+public class MessageBuilder implements GenericMessageBuilder {
     private String messageLocalizationKey;
     private HashMap<String, String> replacements = new HashMap<>();
 
@@ -24,5 +27,14 @@ public class MessageBuilder {
         String newString = baseString;
         for (Map.Entry<String, String> entry : this.replacements.entrySet()) newString = newString.replace(entry.getKey(), entry.getValue());
         return newString;
+    }
+
+    // So we can always iterate
+    public List<MessageBuilder> getMessageBuilders() {
+        return List.of(this);
+    }
+
+    public boolean isMultiBuilder() {
+        return false;
     }
 }
