+++ src/main/java/factionsplusplus/builders/interfaces/GenericMessageBuilder.java
@@ -1,0 +1,10 @@
+package factionsplusplus.builders.interfaces;
+
+import java.util.List;
+
+import factionsplusplus.builders.MessageBuilder;
+
+public interface GenericMessageBuilder {
+    boolean isMultiBuilder();
+    List<MessageBuilder> getMessageBuilders();
+}
