+++ src/main/java/factionsplusplus/commands/ListCommand.java
@@ -7,7 +7,6 @@ package factionsplusplus.commands;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 
-import factionsplusplus.data.PersistentData;
 import factionsplusplus.models.Command;
 import factionsplusplus.models.CommandContext;
 import factionsplusplus.models.Faction;
@@ -16,6 +15,7 @@ import factionsplusplus.services.FactionService;
 import factionsplusplus.builders.CommandBuilder;
 import org.bukkit.ChatColor;
 
+import java.util.Collection;
 import java.util.List;
 
 /**
@@ -24,13 +24,11 @@ import java.util.List;
 @Singleton
 public class ListCommand extends Command {
 
-    private final PersistentData persistentData;
     private final FactionService factionService;
     private final DataService dataService;
 
     @Inject
     public ListCommand(
-        PersistentData persistentData,
         FactionService factionService,
         DataService dataService
     ) {
@@ -41,7 +39,6 @@ public class ListCommand extends Command {
                 .withDescription("List all factions on the server.")
                 .requiresPermissions("mf.list")
         );
-        this.persistentData = persistentData;
         this.factionService = factionService;
         this.dataService = dataService;
     }
@@ -52,14 +49,14 @@ public class ListCommand extends Command {
             return;
         }
         context.replyWith("FactionsTitle");
-        List<PersistentData.SortableFaction> sortedFactionList = this.persistentData.getSortedListOfFactions();
+        Collection<Faction> sortedFactionList = this.factionService.getFactionsByPower();
         context.replyWith("ListLegend");
         context.reply(ChatColor.AQUA + "-----");
-        for (PersistentData.SortableFaction sortableFaction : sortedFactionList) {
-            final Faction temp = sortableFaction.getFaction();
-            context.reply(ChatColor.AQUA + String.format("%-25s %10s %10s %10s", temp.getName(), "P: " +
-                    this.factionService.getCumulativePowerLevel(temp), "M: " + temp.getPopulation(), "L: " +
-                    this.dataService.getClaimedChunksForFaction(temp).size()));
+        for (Faction faction : sortedFactionList) {
+
+            context.reply(ChatColor.AQUA + String.format("%-25s %10s %10s %10s", faction.getName(), "P: " +
+                    this.factionService.getCumulativePowerLevel(faction), "M: " + faction.getPopulation(), "L: " +
+                    this.dataService.getClaimedChunksForFaction(faction).size()));
         }
     }
 }
