+++ src/main/java/factionsplusplus/services/FactionService.java
@@ -3,6 +3,12 @@ package factionsplusplus.services;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 
+import factionsplusplus.utils.Pair;
+import factionsplusplus.utils.StringUtils;
+import factionsplusplus.utils.Comparators;
+import factionsplusplus.builders.MessageBuilder;
+import factionsplusplus.builders.MultiMessageBuilder;
+import factionsplusplus.builders.interfaces.GenericMessageBuilder;
 import factionsplusplus.constants.FlagType;
 import factionsplusplus.models.Faction;
 import factionsplusplus.models.FactionFlag;
@@ -17,7 +23,11 @@ import factionsplusplus.services.DynmapIntegrationService;
 import javax.inject.Provider;
 import java.util.HashMap;
 import java.util.Iterator;
+import java.util.List;
 import java.util.UUID;
+import java.util.Collection;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import java.util.Map;
 
 @Singleton
@@ -196,6 +206,7 @@ public class FactionService {
             faction.removeVassal(targetFaction.getID());
         }
     }
+
     public void removeAllClaimedChunks(Faction faction) {
         Iterator<ClaimedChunk> itr = this.claimedChunkRepository.all().iterator();
         while (itr.hasNext()) {
@@ -203,6 +214,7 @@ public class FactionService {
             if (currentChunk.getHolder().equals(faction.getID())) itr.remove();
         }
     }
+
     public void removeAllOwnedLocks(Faction faction) {
         Iterator<LockedBlock> itr = this.lockedBlockRepository.all().iterator();
 
@@ -212,4 +224,90 @@ public class FactionService {
         }
     }
 
+    public void disbandAllZeroPowerFactions() {
+        this.factionRepository.all().values()
+            .stream()
+            .filter(faction -> this.getCumulativePowerLevel(faction) == 0)
+            .forEach(faction -> {
+                // TODO: send "AlertDisbandmentDueToZeroPower" in some way to the faction
+                this.removeFaction(faction);
+            });
+    }
+
+    public long removeLiegeAndVassalReferencesToFaction(UUID factionUUID) {
+        long changes = this.factionRepository.all().values().stream()
+                .filter(f -> f.isLiege(factionUUID) || f.isVassal(factionUUID))
+                .count(); // Count changes
+
+        this.factionRepository.all().values().stream().filter(f -> f.isLiege(factionUUID)).forEach(f -> f.setLiege(null));
+        this.factionRepository.all().values().stream().filter(f -> f.isVassal(factionUUID)).forEach(Faction::clearVassals);
+
+        return changes;
+    }
+
+    public Collection<Faction> getFactionsByPower() {
+        return this.factionRepository.all().values()
+            .stream()
+            .map(faction -> Pair.of(faction, this.getCumulativePowerLevel(faction)))
+            .sorted(Comparators.FACTIONS_BY_POWER)
+            .map(pair -> pair.left())
+            .collect(Collectors.toList());
+    }
+
+    public List<Faction> getFactionsFromUUIDs(List<UUID> factionUUIDs) {
+        return factionUUIDs.stream()
+            .map(id -> this.factionRepository.get(id))
+            .collect(Collectors.toList());
+    }
+
+    public String getCommaSeparatedFactionNames(List<UUID> factionUUIDs) {
+        return this.getFactionsFromUUIDs(factionUUIDs)
+            .stream()
+            .map(Faction::toString)
+            .collect(Collectors.joining(", "));
+    }
+
+    public GenericMessageBuilder generateFactionInfo(Faction faction) {
+        MultiMessageBuilder builder = new MultiMessageBuilder();
+        // Faction header
+        builder.add(new MessageBuilder("FactionInfo.Title"));
+        // Faction name
+        builder.add(new MessageBuilder("FactionInfo.Name").with("name", faction.getName()));
+        // Owner
+        builder.add(new MessageBuilder("FactionInfo.Owner").with("owner", StringUtils.parseAsPlayer(faction.getOwner()).getName()));
+        // Description (if applicable)
+        if (faction.getDescription() != null) builder.add(new MessageBuilder("FactionInfo.Description").with("desc", faction.getDescription()));
+        // Population
+        builder.add(new MessageBuilder("FactionInfo.Population").with("amount", String.valueOf(faction.getPopulation())));
+        // Allies (if applicable)
+        if (!faction.getAllies().isEmpty()) builder.add(new MessageBuilder("FactionInfo.Allies").with("factions", String.join(", ", this.getCommaSeparatedFactionNames(faction.getAllies()))));
+        // Enemies (if applicable)
+        if (!faction.getEnemyFactions().isEmpty()) builder.add(new MessageBuilder("FactionInfo.AtWarWith").with("factions", String.join(", ", this.getCommaSeparatedFactionNames(faction.getEnemyFactions()))));
+        // Power level
+        final int claimedChunks = this.claimedChunkRepository.getAllForFaction(faction).size();
+        final int cumulativePowerLevel = this.getCumulativePowerLevel(faction);
+        builder.add(new MessageBuilder("FactionInfo.PowerLevel").with("level", String.valueOf(cumulativePowerLevel)).with("max", String.valueOf(this.getMaximumCumulativePowerLevel(faction))));
+        // Demesne Size
+        builder.add(new MessageBuilder("FactionInfo.DemesneSize").with("number", String.valueOf(claimedChunks)).with("max", String.valueOf(cumulativePowerLevel)));
+
+        // Bonus power enabled?
+        if (faction.getBonusPower() != 0) builder.add(new MessageBuilder("BonusPower").with("amount", String.valueOf(faction.getBonusPower())));
+
+        // Is Vassal?
+        if (faction.hasLiege()) {
+            Faction liege = this.factionRepository.get(faction.getLiege());
+            if (liege != null) builder.add(new MessageBuilder("Liege").with("name", liege.getName()));
+        }
+
+        // Is Liege?
+        if (faction.isLiege()) {
+            int vassalContribution = this.calculateCumulativePowerLevelWithVassalContribution(faction) - this.calculateCumulativePowerLevelWithoutVassalContribution(faction);
+            if (this.isWeakened(faction)) vassalContribution = 0;
+            builder.add(new MessageBuilder("Vassals").with("name", this.getCommaSeparatedFactionNames(faction.getVassals())));
+            builder.add(new MessageBuilder("VassalContribution").with("amount", String.valueOf(vassalContribution)));
+        }
+
+        // Send off!
+        return builder;
+    }
 }
