+++ src/main/java/factionsplusplus/utils/Comparators.java
@@ -1,0 +1,13 @@
+package factionsplusplus.utils;
+
+import java.util.Comparator;
+import java.util.Map;
+
+import factionsplusplus.models.Faction;
+
+public class Comparators {
+    public static final Comparator<Pair<Faction, Integer>> FACTIONS_BY_POWER = (pairOne, pairTwo) -> {
+        int comparison = Integer.compare(pairOne.right(), pairTwo.right()); 
+        return Integer.compare(0, comparison);
+    };
+}
