+++ src/main/resources/language.yml
@@ -280,15 +280,19 @@ CannotAttackNonWarringPlayer: "&cYou can't attack another player if your faction
 ConfigListPage: "&a===== Config List Page #page#/#pages# ====="
 UsageConfigShow: "&cUsage: /mf config show <page>"
 FactionInfo:
-  - "&b#faction# Faction Info"
-  - "&bName: #name#"
-  - "&bOwner: #owner#"
-  - "&bDescription: #desc#"
-  - "&bPopulation: #pplt#"
-  - "&bAllied With: #aw#"
-  - "&bAt War With: #aww#"
-  - "&bPower Level: #pl#/#pl_max#"
-  - "&bDemesne Size: #number#/#max#"
+  Title: "&bFaction Info"
+  Name: "&bName: #name#"
+  Owner: "&bOwner: #owner#"
+  Description: "&bDescription: #desc#"
+  Population: "&bPopulation: #amount#"
+  Allies: "&bAllied With: #factions#"
+  AtWarWith: "&bAt War With: #factions#"
+  PowerLevel: "&bPower Level: #level#/#max#"
+  DemesneSize: "&bDemesne Size: #number#/#max#"
+  BonusPower: "&aBonus Power From Admins: #amount#"
+  VassalContribution: "&aBonus Power From Vassals: #amount#"
+  Liege: "&bLiege: #name#"
+  Vassals: "&bVassals: #name#"
 PluginInfo:
   - "&b== Factions Plus Plus #version# =="
   - "&bDevelopers: #dev#"
